Add footer storybook component
